+++ packages/parsers/src/index.ts
@@ -2,3 +2,6 @@ export * from "./detector";
 export * from "./junit-xml";
 export * from "./selenium-xml";
 export * from "./appium-xml";
+export * from "./allure-zip";
+export * from "./intellij-log";
+export * from "./appium-log";